Split wrap_qml to get around mingw size limits
Useful for debug build
